+++ lib/src/utils/utils.dart
@@ -1,14 +1,17 @@
 import 'dart:math';
 
 class Utils {
-  final random = Random();
 
-  String get generateId => 'el-id-${random.nextInt(99999999)}-${random.nextInt(99999999)}';
+  String generateId(){
+    return 'el-id-${Random.secure().nextInt(99999999)}-${Random.secure().nextInt(99999999)}';
+  }
 
-  String relatedJs({required String id}) => """var mainModelViewer = document.querySelector('#$id');
-cameraOrbit = (a, b, c) => {console.log("------camOrbiit------"); mainModelViewer.cameraOrbit = `\${a}deg \${b}deg \${c}m`; }
-cameraTarget = (x, y, z) => {mainModelViewer.cameraTarget = `\${x}m \${y}m \${z}m`}
-customEvaluate = (code) => { eval(code) }
-""";
+  String injectedJS(){
+    String jsStr = """
+    customEvaluate = (code) => { eval(code) }
+    customEvaluateWithResult = (code) => { return eval (code) }
+    """;
+    return jsStr;
+  }
 
 }